Backend/requirements.txt
appnope==0.1.3
asgiref==3.3.4
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
decorator==5.1.1
Django==2.2.19
django-cors-headers==3.7.0
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
executing==0.8.3
ipython==8.3.0
jedi==0.18.1
jwt==1.3.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
